refactor(modal): openConfirmPopup을 활용하여 confirm popup 리팩터링

+++ features/ticket/delete-ticket/ui/TicketDeleteButton.tsx
@@ -1,38 +1,41 @@
-import {useCommonSlice} from '@/slice/commonSlice';
+import { useCommonSlice, usePopup } from '@/slice/commonSlice';
 import React from 'react';
-import {TouchableOpacity} from 'react-native';
-import {useDeleteTicket} from '../model';
-import {Txt} from '@/shared/ui';
-import {color_token} from '@/constants/theme';
+import { TouchableOpacity } from 'react-native';
+import { useDeleteTicket } from '../model';
+import { Txt } from '@/shared/ui';
+import { color_token } from '@/constants/theme';
 
 const TicketDeleteButton = ({ticketId}: {ticketId: number}) => {
   const {modal} = useCommonSlice();
+
+  const {openConfirmPopup} = usePopup()
+
   const {deleteTicket} = useDeleteTicket();
 
   const onDeleteTicket = () => {
-    modal.open({
-      header: '안내',
+    openConfirmPopup({
       content: '해당 티켓을 삭제할까요?',
-      button: [
-        {
-          text: '취소',
-          onPress: modal.hide,
-          buttonStyle: {
-            backgroundColor: color_token.gray300,
-          },
+      confirm: {
+        text: '삭제',
+        onPress: () => deleteTicket(ticketId),
+        buttonStyle: {
+          backgroundColor: color_token.primary,
         },
-        {
-          text: '삭제',
-          onPress: () => deleteTicket(ticketId),
-          buttonStyle: {
-            backgroundColor: color_token.primary,
-          },
-          buttonTextStyle: {
-            color: color_token.white,
-          },
+        buttonTextStyle: {
+          color: color_token.white,
         },
-      ],
-    });
+      },
+      cancel: {
+        text: '취소',
+        onPress: modal.hide,
+        buttonStyle: {
+          backgroundColor: color_token.gray300,
+        },
+        buttonTextStyle: {
+          color: color_token.gray900,
+        },
+      },
+    })
   };
 
   return (
